deployment/09-a2a/pyproject.toml
[project]
name = "aws-assistant-strands"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "termcolor",
    "nest_asyncio",
    "a2a-sdk>=0.2.1a1",
    "mcp[cli]>=1.9.0",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
]
